Escape unique id selector
