+++ app/electron/main/code/diff/class.ts
@@ -1,0 +1,44 @@
+import traverse from '@babel/traverse';
+import t from '@babel/types';
+import { twMerge } from 'tailwind-merge';
+
+export function addClassToAst(ast: t.File, className: string) {
+    let processed = false;
+    traverse(ast, {
+        JSXOpeningElement(path) {
+            if (processed) {
+                return;
+            }
+
+            let classNameAttr = null;
+
+            path.node.attributes.forEach((attribute) => {
+                if (t.isJSXAttribute(attribute) && attribute.name.name === 'className') {
+                    classNameAttr = attribute;
+
+                    if (t.isStringLiteral(attribute.value)) {
+                        attribute.value.value = twMerge(attribute.value.value, className);
+                    }
+                    // Handle className that is an expression (e.g., cn("class1", className))
+                    else if (
+                        t.isJSXExpressionContainer(attribute.value) &&
+                        t.isCallExpression(attribute.value.expression)
+                    ) {
+                        attribute.value.expression.arguments.push(t.stringLiteral(className));
+                    }
+                }
+            });
+
+            if (!classNameAttr) {
+                const newClassNameAttr = t.jsxAttribute(
+                    t.jsxIdentifier('className'),
+                    t.stringLiteral(className),
+                );
+                path.node.attributes.push(newClassNameAttr);
+            }
+
+            path.stop();
+            processed = true;
+        },
+    });
+}
