+++ app/electron/main/code/diff/insert.ts
