+++ app/electron/main/code/helpers.ts
@@ -1,0 +1,22 @@
+import { parse } from '@babel/parser';
+import t from '@babel/types';
+
+export function removeSemiColonIfApplicable(code: string, original: string) {
+    if (!original.endsWith(';') && code.endsWith(';')) {
+        return code.slice(0, -1);
+    }
+    return code;
+}
+
+export function parseJsx(code: string): t.File | undefined {
+    try {
+        return parse(code, {
+            plugins: ['typescript', 'jsx'],
+            sourceType: 'module',
+            allowImportExportEverywhere: true,
+        });
+    } catch (e) {
+        console.error('Error parsing code', e);
+        return;
+    }
+}
